refactor: merge ColonyTabs component into ColonyPage
